Moved width-finding functions to library
