+++ frontend/packages/sheet/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@blueprintjs/core": "^3.39.0",
     "@macrostrat/hyper": "^1.2.12",
+    "d3-array": "^2.12.1",
     "react": "^17.0.2",
     "react-dnd": "^14.0.2",
     "react-dnd-html5-backend": "^14.0.0",
