+++ frontend/packages/sheet/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./provider";
 export * from "./components";
 export * from "./helpers";
+export * from "./columns";
